# Documentation/api/openapi/weatherwear-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeatherWear API",
    "description": "REST API for user authentication, weather lookup, AI clothing recommendations, recommendation history, feedback, and an AI style chat assistant.",
    "version": "1.0.0",
    "contact": {
      "name": "Dubrovin Mikita",
      "email": "dubrovin.nikita@student.ehu.lt"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090/api",
      "description": "Local development server"
    },
    {
      "url": "https://wwtodipl-production.up.railway.app/api",
      "description": "Production server configured by application-prod.yml"
    }
  ],
...
  "paths": {
...
    "/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "operationId": "sendChatMessage",
        "summary": "Send a chat message",
        "description": "Sends a message to the AI style assistant. If sessionId is omitted, a new chat session is created. If city is supplied, the prompt includes current weather for that city.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              },
              "examples": {
                "newSession": {
                  "summary": "Start a new session",
                  "value": {
                    "message": "What should I wear for a long walk today?",
                    "city": "Vilnius"
                  }
                },
                "existingSession": {
                  "summary": "Continue an existing session",
                  "value": {
                    "sessionId": 12,
                    "message": "Can you make it more business casual?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assistant response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                },
                "examples": {
                  "answer": {
                    "value": {
                      "sessionId": 12,
                      "answer": "Choose a breathable base layer, a light jacket, comfortable trousers, and shoes with grip. Carry a compact umbrella if rain is likely.",
                      "createdAt": "2026-04-30T12:45:00"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGatewayError"
          }
        }
      }
    },
    "/chat/sessions": {
      "get": {
        "tags": [
          "Chat"
        ],
        "operationId": "getChatSessions",
        "summary": "List current user's chat sessions",
        "description": "Returns chat sessions ordered by most recent update.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Chat sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatSessionResponse"
                  }
                },
                "examples": {
                  "sessions": {
                    "summary": "Chat session list",
                    "value": [
                      {
                        "id": 12,
                        "title": "Style chat",
                        "createdAt": "2026-04-30T12:40:00",
                        "updatedAt": "2026-04-30T12:45:00"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/chat/sessions/{id}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "operationId": "getChatSessionMessages",
        "summary": "Get messages from a chat session",
        "description": "Returns messages from one chat session if it belongs to the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Chat session identifier.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 12
          }
        ],
        "responses": {
          "200": {
            "description": "Messages in chronological order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageDto"
                  }
                },
                "examples": {
                  "messages": {
                    "summary": "Chat session messages",
                    "value": [
                      {
                        "id": 9001,
                        "role": "USER",
                        "content": "What should I wear for a long walk today?",
                        "createdAt": "2026-04-30T12:44:00"
                      },
                      {
                        "id": 9002,
                        "role": "ASSISTANT",
                        "content": "Choose a breathable base layer, a light jacket, comfortable trousers, and shoes with grip.",
                        "createdAt": "2026-04-30T12:45:00"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Chat"
        ],
        "operationId": "deleteChatSession",
        "summary": "Delete a chat session",
        "description": "Deletes a chat session if it belongs to the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Chat session identifier.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 12
          }
        ],
        "responses": {
          "204": {
            "description": "Chat session deleted successfully. No response body is returned."
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned by /auth/register or /auth/login. Send it as Authorization: Bearer <token>."
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed, a required parameter is missing, or the JSON body is malformed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validation": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "email: Email must be valid",
                  "path": "/api/auth/register"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Authentication is missing, invalid, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Authentication is required",
                  "path": "/api/preferences"
                }
              }
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "The requested resource was not found or does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "notFound": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Recommendation history not found",
                  "path": "/api/history/999"
                }
              }
            }
          }
        }
      },
...
      "BadGatewayError": {
        "description": "An upstream dependency such as OpenWeather or the LLM provider failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "upstream": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 502,
                  "error": "Bad Gateway",
                  "message": "Failed to fetch weather data",
                  "path": "/api/weather"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "serverError": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Unexpected server error",
                  "path": "/api/recommendations"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int64",
            "description": "Existing chat session ID. Omit to start a new session.",
            "example": 12
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 3000,
            "example": "What should I wear for a long walk today?"
          },
          "city": {
            "type": "string",
            "maxLength": 120,
            "description": "Optional city used to add current weather context to the prompt.",
            "example": "Vilnius"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int64",
            "example": 12
          },
          "answer": {
            "type": "string",
            "example": "Choose a breathable base layer, a light jacket, comfortable trousers, and shoes with grip."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          }
        }
      },
      "ChatSessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 12
          },
          "title": {
            "type": "string",
            "example": "Style chat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:40:00"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          }
        }
      },
      "ChatMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 9001
          },
          "role": {
            "$ref": "#/components/schemas/ChatRole"
          },
          "content": {
            "type": "string",
            "example": "What should I wear for a long walk today?"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "timestamp",
          "status",
          "error",
          "message",
          "path"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "email: Email must be valid"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register"
          }
        }
      },
...
      "ChatRole": {
        "type": "string",
        "enum": [
          "USER",
          "ASSISTANT",
          "SYSTEM"
        ],
        "example": "ASSISTANT"
      }
    }
  }
}